requirements.txt
pyautogui
Pillow
numpy
google-generativeai
